# Speed up outputs and fetching blobs