--- a/crates/brioche-core/src/output.rs
+++ b/crates/brioche-core/src/output.rs
@@ -429,7 +429,7 @@ async fn fetch_descendent_artifact_blobs(
     crate::references::descendent_artifact_blobs(brioche, [artifact.clone()], &mut blobs).await?;
 
     // Fetch all referenced blobs
-    crate::registry::fetch_blobs(brioche.clone(), &blobs).await?;
+    crate::registry::fetch_blobs(brioche.clone(), blobs).await?;
 
     Ok(())
 }