This PR makes two quick and easy optimizations, mostly affecting cache hits:

1. Check for existing blobs in `fetch_blobs` using a `tokio::task::spawn_blocking` task, rather than running lots of futures concurrently. Since Tokio filesystem I/O is inherently blocking, trying to split out the work into more parallelism ended up creating a lot of overhead, so just doing the checks in a single blocking function turned out to be _a lot_ faster
2. Update `create_local_output` to only fetch descendent blobs if the local output doesn't already exist. If the local output already exists, we can short-circuit sooner, which turns out to be a big speedup

For testing, I was trying builds against `xplr` from this commit: https://github.com/brioche-dev/brioche-packages/commit/ff64f2d8edc4e814c0c419aae5f497ddc6f7c359. The build (expectedly) fails each iteration, but since the process's inputs were supposed to be fast cache hits, it was helpful to see why it was taking so long on each attempt.

Here's a few screenshots of Jaeger before and after this PR (using the changes from #168 to make the output more readable):

Before (15.51s):
![Screenshot of Jaeger, showing a Brioche build that took 15.51 seconds. The bulk of the time is taken by a call to "bake_process", which spawns a waterfall of "fetch_descendent_blobs" calls that go beyond the edge of the screen](https://github.com/user-attachments/assets/d4aebde2-f02f-47ae-9193-dc7711883018)


After (1.12s):
![Screenshot of Jaeger, showing a Brioche build that took 1.12 seconds. 689ms is spent in the "evaluate" function, and "bake" takes 354ms. It looks very similar to the previous screenshot, but without the cascade of "fetch_descendent_blobs" calls](https://github.com/user-attachments/assets/37db2ab3-8d90-47f3-8f20-043d1429cada)